pyproject.toml
[project]
name = "apx"
dynamic = ["version"]
description = "apx is the toolkit for building Databricks Apps"
readme = { file = "README.md", content-type = "text/markdown" }
authors = [{ name = "renardeinside", email = "polarpersonal@gmail.com" }]
requires-python = ">=3.11"
dependencies = [
    "basedpyright>=1.32.1",
    "databricks-sdk>=0.68.0",
    "fastapi>=0.119.0",
    "hatchling>=1.27.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "keyring>=25.6.0",
    "mcp>=1.0.0",
    "psutil>=6.1.0",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
    "tenacity>=9.1.2",
    "typer>=0.19.2",
    "typing-extensions>=4.12.0",
    "uvicorn>=0.37.0",
    "watchfiles>=1.1.1",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Framework :: Hatch",
]

[build-system]
requires = ["hatchling", "uv-dynamic-versioning>=0.7.0"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
format-jinja = "{{version}}"

[dependency-groups]
dev = [
    "basedpyright>=1.32.1",
    "databricks-sdk>=0.68.0",
    "fastapi>=0.119.0",
    "mypy>=1.18.2",
    "pydantic-settings>=2.11.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
    "ruff>=0.14.0",
    "sqlmodel>=0.0.27",
    "types-cachetools>=6.2.0.20251022",
    "types-cffi>=1.17.0.20250915",
    "types-greenlet>=3.2.0.20250915",
    "types-jinja2>=2.11.9",
    "types-pexpect>=4.9.0.20250916",
    "types-protobuf>=6.32.1.20250918",
    "types-psutil>=7.0.0.20251001",
    "types-pyasn1>=0.6.0.20250914",
    "types-pygments>=2.19.0.20250809",
    "types-pyyaml>=6.0.12.20250915",
    "types-requests>=2.32.4.20250913",
    "types-ujson>=5.10.0.20250822",
]

[project.scripts]
apx = "apx.__main__:entrypoint"

[tool.hatch.build.hooks.version]
path = "src/apx/_version.py"
template = '''
version = "{version}"
'''

[tool.basedpyright]
exclude = ["src/apx/templates", "**/node_modules", "**/__pycache__", "**/.*"]
reportUnusedCallResult = false
reportCallInDefaultInitializer = false
reportUnusedFunction = false


[tool.mypy]
exclude = ['venv', '.venv', 'src/apx/templates', 'node_modules']

[tool.coverage.run]
parallel = true
source = ["src/apx"]
omit = [
    "tests",
    "src/apx/templates",
    "**/node_modules",
    "**/__pycache__",
    "**/.*",
    "**/.jinja2"
]
